requirements.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
dill==0.3.8
Django==5.0.9
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
executing==1.2.0
gunicorn==25.0.0
iniconfig==2.3.0
ipython==8.7.0
isort==5.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mssql-django==1.5
mysqlclient==2.2.7
packaging==26.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
pluggy==1.6.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.11.0
pyodbc==5.2.0
pytest==9.0.2
pytest-django==4.11.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.2
tomli==2.4.0
tomlkit==0.12.3
traitlets==5.8.0
typing_extensions==4.15.0
tzdata==2024.2
watermark==2.3.1
wcwidth==0.2.5
